+++ src/user-icon.ts
@@ -11,31 +11,33 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
 
 // @ts-check
 
-import { html, LitElement, css }        from 'lit-element';
+import { html, LitElement, css, customElement, property }        from 'lit-element';
 import { connect }                      from 'pwa-helpers/connect-mixin.js';
-import { store }                        from './store';
+import { store, RootState }             from './store';
 import { userStyles }                   from './styles';
 import { openSign, setImage }           from './user-action';
 import { profileURL }                   from './user-functions';
 
 export const faceIcon = html`<svg viewBox="0 0 24 24" height="32px" width="32px"><path d="M9 11.75c-.69 0-1.25.56-1.25 1.25s.56 1.25 1.25 1.25 1.25-.56 1.25-1.25-.56-1.25-1.25-1.25zm6 0c-.69 0-1.25.56-1.25 1.25s.56 1.25 1.25 1.25 1.25-.56 1.25-1.25-.56-1.25-1.25-1.25zM12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 18c-4.41 0-8-3.59-8-8 0-.29.02-.58.05-.86 2.36-1.05 4.23-2.98 5.21-5.37C11.07 8.33 14.05 10 17.42 10c.78 0 1.53-.09 2.25-.26.21.71.33 1.47.33 2.26 0 4.41-3.59 8-8 8z"></path></svg>`;
 
+@customElement('user-icon')
 export class UserIcon extends connect(store)(LitElement) {
 
-  static get properties() {
-    return {
-      _user:            { type: String },
-      _photoURL:        { type: String }
-    };
-  }
+  @property({type: Boolean})
+  private _user = false;
+
+  @property({type: String})
+  private _photoURL = '';
+
 
   constructor() {
     super();
   }
 
-  firstUpdated() {
-    this.shadowRoot.getElementById('subscribe').addEventListener('click', () => { store.dispatch(openSign()) } );
+  protected firstUpdated() {
+    this.shadowRoot!.getElementById('subscribe').addEventListener('click', () => { store.dispatch(openSign()) } );
   
+    // @ts-ignore
     firebase.auth().onAuthStateChanged( (firebaseUser) => {
       if (firebaseUser) { /* INCLUDE */ this._photoURL = profileURL(); }
       else              { /* EXCLUDE */ }
@@ -43,9 +45,9 @@ export class UserIcon extends connect(store)(LitElement) {
     });
   }
   
-  stateChanged(state) {
-    this._user        = state.user.currentUser;
-    this._photoURL    = state.user.photoURL;
+  stateChanged(state: RootState) {
+    this._user        = state.user!.currentUser;
+    this._photoURL    = state.user!.photoURL;
   }  
 
 
@@ -79,7 +81,7 @@ export class UserIcon extends connect(store)(LitElement) {
         `
     ]}
   
-  render() {
+  protected render() {
     return html`
 
       <button id="subscribe" aria-label="login">
@@ -102,4 +104,3 @@ export class UserIcon extends connect(store)(LitElement) {
     `;
   }
 }
-window.customElements.define('user-icon', UserIcon);